Moved first js files to ts
