+++ src/package.json
@@ -87,6 +87,7 @@
     "@types/express": "^4.17.21",
     "@types/formidable": "^3.4.5",
     "@types/http-errors": "^2.0.4",
+    "@types/jquery": "^3.5.30",
     "@types/jsdom": "^21.1.7",
     "@types/jsonwebtoken": "^9.0.6",
     "@types/mocha": "^10.0.7",
